pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pixels_to_predictions"]

[project]
name = "pixels-to-predictions"
version = "0.1.0"
description = "Multimodal MCQ: SmolVLM-500M-Instruct fine-tune for NYU DL Spring 2026 Kaggle final"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.2.0",
    "pillow>=10.4.0",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.17.0",
    "ruff>=0.12.0",
    "pandas-stubs>=2.2.0",
    "types-requests>=2.32.0",
    "types-tqdm>=4.67.0",
    "types-pyyaml>=6.0.0",
]
# Heavy ML deps (torch/transformers/peft) are version-pinned in a separate
# extra because the correct wheel depends on hardware:
#   - x86_64 CUDA 12.x (Colab / Kaggle)          -> `uv pip install -e .[gpu-x86]`
#   - aarch64 CUDA 13 on Jetson Thor (Advantech) -> `uv pip install -e .[gpu-thor]`
gpu-x86 = [
    "torch>=2.5.0",
    "transformers>=4.48.0",
    "peft>=0.14.0",
    "trl>=0.14.0",
    "accelerate>=1.3.0",
    "bitsandbytes>=0.45.0",
    "datasets>=3.2.0",
    "num2words>=0.5.0",
]
gpu-thor = [
    # aarch64 wheels for Jetson Thor; installed via extra-index-url on the box.
    "transformers>=4.48.0",
    "peft>=0.14.0",
    "trl>=0.14.0",
    "accelerate>=1.3.0",
    "datasets>=3.2.0",
]
search = [
    "matplotlib>=3.9.0",
    "numpy>=2.0.0",
    "pyyaml>=6.0.0",
]

[project.scripts]
p2p-search = "pixels_to_predictions.search.__main__:main"

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "D203",    # incompatible with D211 (one-blank-line-before-class)
    "D213",    # incompatible with D212 (multi-line-summary-first-line)
    "S101",    # assert is fine in research code
    "T201",    # print is fine for CLI/training scripts
    "FBT001",  # boolean positional args -- common in ML configs
    "FBT002",  # boolean defaults -- common in ML configs
    "ERA001",  # commented-out code -- useful during experimentation
    "PLR0913", # too many args -- ML training functions have many knobs
    "PLR2004", # magic numbers -- hyperparameters appear inline
    "PLC0415", # deferred imports -- needed to delay torch/transformers until actually used
    "COM812",  # trailing comma -- conflicts with ruff format
    "ISC001",  # single-line implicit string concat -- conflicts with ruff format
    "TD002",   # todo author -- solo project
    "TD003",   # todo link -- solo project
    "FIX002",  # todos are fine during development
    "D102",    # public-method docstrings -- too noisy in dataclass-heavy ML code
    "D105",    # magic-method docstrings -- __repr__/__str__ rarely need them
    "S311",    # non-crypto random -- exactly what hyperparam sampling uses
    "PLR0915", # too-many-statements -- report generators and CLI dispatchers have many
    "ANN401",  # typing.Any -- required in generic Protocols and unstructured config dicts
    "TID252",  # relative imports inside the package -- intentional
    "FBT003",  # boolean positional values in external API calls
    "UP042",   # str+Enum idiom is deliberate (JSON-friendly string values)
]

[tool.ruff.lint.per-file-ignores]
"references/*" = ["ALL"]
"stubs/**/*.pyi" = ["ANN401", "S105", "PLR0913", "D"]
"scripts/*.py"  = ["D100", "D103", "INP001"]  # scripts don't need a package docstring
"configs/**/*.py" = ["INP001"]  # search-space configs are imported by path, not as a package
"**/__main__.py" = ["D100"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
strict = true
explicit_package_bases = true
mypy_path = ["src", "stubs"]
warn_unused_ignores = false
show_error_codes = true
show_column_numbers = true

[[tool.mypy.overrides]]
module = [
    "torch.*",
    "transformers.*",
    "peft.*",
    "trl.*",
    "accelerate.*",
    "bitsandbytes.*",
    "datasets.*",
    "num2words.*",
    "unsloth.*",
    "PIL.*",
    "matplotlib.*",
    "yaml.*",
]
ignore_missing_imports = true
